requirements.txt
tinify==1.7.1
Pillow==9.4.0